--- noti-service/src/main/java/com/waither/notiservice/dto/LocationDto.java
@@ -1,16 +1,0 @@
-package com.waither.notiservice.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class LocationDto {
-
-    public double x;
-
-    public double y;
-}
